refactor(ai): apply architecture review — placement and genericity
Act on the placement/genericity review while keeping behavior unchanged:

- Split pkg/gofr/mcp.go: tools.go now owns the generic route->tool exposure
  (routerTools, request synthesis) behind an explicit registerTools step; mcp.go
  owns only the MCP transport. Tool exposure is separated from serving.
- Move the generic recording ResponseWriter to pkg/gofr/http as ResponseCapture,
  next to the other writer wrappers, so it is reusable and out of MCP glue.
- Container stores the instrumented ai.LLM in one field, built in SetLLM; a mock
  container injects through the same field, removing the test-only llmMock field.
  A lazy tools indirection keeps late-registered tools visible; health checks the
  raw provider (llmModel).
- llm.Client reads keys only through the injected config.Config; drop the direct
  os.Getenv fallback (GoFr config is already env-backed).
- Forward auth headers via middleware.AuthHeaders(), the single source of truth
  for which headers auth reads, instead of a hardcoded list in MCP glue.
- Rename MCPExclude to WithExcludedRoutes for a consistent With* option set.
- Document Instrument/StartCall/Recorder as the sanctioned extension point.
